onlineboutique/src/recommendationservice/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
apache-skywalking == 1.0.1
backoff==2.2.1
    # via opentelemetry-exporter-otlp-proto-grpc
cachetools==5.3.0
    # via google-auth
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
deprecated==1.2.13
    # via opentelemetry-api
google-api-core==2.11.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-api-python-client==2.79.0
    # via google-cloud-profiler
google-auth==2.16.1
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-profiler
google-auth-httplib2==0.1.0
    # via
    #   google-api-python-client
    #   google-cloud-profiler
google-cloud-profiler==4.0.0
    # via -r requirements.in
googleapis-common-protos==1.58.0
    # via
    #   google-api-core
    #   opentelemetry-exporter-otlp-proto-grpc
grpcio==1.51.3
    # via
    #   -r requirements.in
    #   grpcio-health-checking
    #   opentelemetry-exporter-otlp-proto-grpc
grpcio-health-checking==1.51.3
    # via -r requirements.in
httplib2==0.21.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via requests
opentelemetry-api==1.16.0
    # via
    #   opentelemetry-distro
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-grpc
    #   opentelemetry-sdk
opentelemetry-distro==0.37b0
    # via -r requirements.in
opentelemetry-exporter-otlp-proto-grpc==1.16.0
    # via -r requirements.in
opentelemetry-instrumentation==0.37b0
    # via
    #   opentelemetry-distro
    #   opentelemetry-instrumentation-grpc
opentelemetry-instrumentation-grpc==0.37b0
    # via -r requirements.in
opentelemetry-proto==1.16.0
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.16.0
    # via
    #   opentelemetry-distro
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-instrumentation-grpc
opentelemetry-semantic-conventions==0.37b0
    # via
    #   opentelemetry-instrumentation-grpc
    #   opentelemetry-sdk
protobuf==4.22.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-profiler
    #   googleapis-common-protos
    #   grpcio-health-checking
    #   opentelemetry-proto
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==3.0.9
    # via httplib2
python-json-logger==2.0.7
    # via -r requirements.in
requests==2.28.2
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-profiler
rsa==4.9
    # via
    #   -r requirements.in
    #   google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
typing-extensions==4.5.0
    # via opentelemetry-sdk
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.14
    # via requests
wrapt==1.14.1
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-grpc

# The following packages are considered to be unsafe in a requirements file:
# setuptools
